web/requirements.txt
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-rest==0.8.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
gunicorn==23.0.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scipy==1.16.3
six==1.17.0
sql==2022.4.0
sqlparse==0.5.3
typing_extensions==4.15.0
